requirements.txt
flask==3.0.0
flask-socketio==5.3.5
flask-cors==4.0.0
python-socketio==5.10.0
python-socketio[client]
eventlet==0.33.3
websocket-client==1.6.4
